working website/backend/requirements.txt
google-generativeai
python-telegram-bot
streamlit
flask-cors
Flask